added authentication to post_create, post_delete, added include/post.html

+++ instagram/post/views.py
@@ -1,6 +1,7 @@
 """
 post_list뷰를 'post/' URL에 할당
 """
+
 from django.http import HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 
@@ -22,13 +23,29 @@ def post_list(request):
     return render(request, 'post/post_list.html', context)
 
 
+def post_detail(request, post_pk):
+    post = get_object_or_404(Post, pk=post_pk)
+
+    comment_form = PostCommentForm()
+    context = {
+        'post': post,
+    }
+    return render(request, 'post/post_detail.html', context)
+
+
 def post_create(request):
     """
     HTTP GET: Post를 생성하는 빈 form을 보여주기
     HTTP POST: Post 생성하기
+
+    1. 이 뷰에 접근할 때 해당 사용자가 인증된 상태가 아니면 로그인 뷰로 redirect
+    2. form.is_valid()를 통과한 후 생성하는 Post객체에 author 정보를 추가하기
     :param request:
     :return:
     """
+    if not request.user.is_authenticated:
+        return redirect('member:login')
+
     if request.method == "POST":
         # POST 요청의 경우 PostForm 인스턴스 생성과정에서 request.POST, request.FILES를 사용
         form = PostForm(request.POST, request.FILES)
@@ -36,6 +53,7 @@ def post_create(request):
         if form.is_valid():
             photo = form.cleaned_data.get('photo')
             Post.objects.create(
+                author=request.user,
                 photo=photo
             )
             return redirect('post:list')
@@ -54,43 +72,46 @@ def post_create(request):
     return render(request, 'post/post_form.html', context)
 
 
-def post_detail(request, post_pk):
-
-    post = get_object_or_404(Post, pk=post_pk)
-
-    comment_form = PostCommentForm()
-    context = {
-        'post': post,
-        'comment_form': comment_form,
-    }
-    return render(request, 'post/post_detail.html', context)
-
-
-def post_delete(request):
+def post_delete(request, pk):
     """
-
+    HTTP POST: Post 삭제하기
     :param request:
     :return:
     """
+    if not request.user.is_authenticated:
+        return redirect('post:list')
+
+    if request.method == "POST":
+        post = Post.objects.get(pk=pk)
+        if request.user == post.author:
+            post.delete()
+
+    return redirect('post:list')
 
 
 def post_comment_create(request, post_pk):
     """
     HTTP POST: PostComment 생성하기
+    로그인한 유저만 요청 가능하도록 함
+    작성하는 Comment에 author정보 추가
+
     :param request:
     :return:
     """
+    if not request.user.is_authenticated:
+        return redirect('member:login')
+
     post = get_object_or_404(Post, pk=post_pk)
     if request.method == "POST":
         form = PostCommentForm(request.POST)
         if form.is_valid():
             instance = form.save(commit=False)
             instance.post = post
-            form.save()
+            instance.author = request.user
+            instance.save()
             next = request.GET.get('next')
             if next:
                 return redirect(next)
-
 
     return redirect('post:list')
 
@@ -101,8 +122,12 @@ def post_comment_delete(request, pk):
     :param request:
     :return:
     """
+    if not request.user.is_authenticated:
+        return redirect('post:list')
+
     if request.method == "POST":
         comment = PostComment.objects.get(pk=pk)
-        comment.delete()
+        if request.user == comment.author:
+            comment.delete()
 
     return redirect('post:list')
